+++ e2e/global-setup.ts
@@ -1,8 +1,0 @@
-import { clerkSetup } from "@clerk/testing/playwright";
-import * as dotenv from "dotenv";
-
-dotenv.config({ path: ".env.local" });
-
-export default async function globalSetup() {
-  await clerkSetup();
-}